requirements.txt
openpyxl>=3.1.0
flask>=3.0.0
openai>=1.0.0
google-genai>=1.0.0
snowflake-connector-python>=3.10.0
cryptography>=42.0.0
